Address review committee feedback (round 1)
- Add mjs to Prettier and lefthook format-staged globs so smoke scripts
  are covered by format:check CI gate
- Extract inline consumer check script from smoke-tarball.mjs to
  scripts/smoke-consumer-check.mjs (lintable, formattable, version-controlled)
- Format smoke-artifact.mjs and smoke-tarball.mjs under the new glob
